% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_distance.R
\name{calculate_max_distance}
\alias{calculate_max_distance}
\title{Helper function for get_max_data()}
\usage{
calculate_max_distance(data)
}
\arguments{
\item{data}{}
}
\value{
\code{list} of 2, an index and a value
}
\description{
Helper function for get_max_data()
}
