% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_distance.R
\name{get_max_data}
\alias{get_max_data}
\title{Calculates the distance between each consecutive points
Ignores if the time difference is more than 10 mins}
\usage{
get_max_data(data, ship_name, time_difference = 10)
}
\arguments{
\item{ship_name}{\code{character}}

\item{time_difference}{\code{numeric}}

\item{r}{\code{reactiveValues}}
}
\value{
\code{data.frame} containing a single row
}
\description{
Calculates the distance between each consecutive points
Ignores if the time difference is more than 10 mins
}
